Use data context for the DriverBox control
