+++ DriversBackup/Styles/General.xaml
@@ -1,6 +1,8 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:converters="clr-namespace:DriversBackup.Converters">
+                    xmlns:local="clr-namespace:DriversBackup.Controls"
+                    xmlns:converters="clr-namespace:DriversBackup.Converters"
+                    xmlns:models="clr-namespace:DriversBackup.Models">
 
     <!-- Colors -->
     <SolidColorBrush x:Key="OrangeAccent" Color="#FF5325"/>
@@ -11,6 +13,7 @@
     
     <!-- Converters -->
     <converters:ButtonTypeToColorConverter x:Key="ButtonTypeToColorConverter"/>
+   
 
     <!-- Texts -->
     <Style TargetType="TextBlock" x:Key="PageHeader">
@@ -325,7 +328,7 @@
                             <Setter Property="BorderThickness" Value="0,0,0,0"/>
                         </Trigger>
                         <Trigger Property="IsMouseOver" Value="true">
-                           <Setter Property="Cursor" Value="Hand"/>
+                            <Setter Property="Cursor" Value="Hand"/>
                         </Trigger>
                         <Trigger Property="IsPressed" Value="true">
                             <Setter Property="Opacity" Value="0.9"/>
