+++ DriversBackup/Views/Controls/DriversBox.xaml
@@ -1,12 +1,11 @@
-﻿<UserControl x:Class="DriversBackup.Controls.DriversBox"
+﻿<UserControl x:Class="DriversBackup.Views.Controls.DriversBox"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:DriversBackup.Controls"
-             xmlns:vm="clr-namespace:DriversBackup.ViewModels"
              xmlns:p="clr-namespace:DriversBackup.Properties"
              xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+             xmlns:controls="clr-namespace:DriversBackup.Controls"
              mc:Ignorable="d" Background="White"
              d:DesignHeight="300" d:DesignWidth="550">
     <Grid>
@@ -17,7 +16,7 @@
         </Grid.RowDefinitions>
         <!--Search textbox -->
         <Grid>
-            <local:PlaceholderTextBox x:Name="SearchBox"
+            <controls:PlaceholderTextBox x:Name="SearchBox"
                                       Text="{Binding Search, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
                                       PlaceholderText="{x:Static p:Resources.SearchDots}"
                                       Padding="5,0,0,0" FontSize="20" VerticalContentAlignment="Center"
@@ -28,7 +27,7 @@
                         <i:InvokeCommandAction Command="{Binding SortByCommand}" CommandParameter="Search" />
                     </i:EventTrigger>
                 </i:Interaction.Triggers>
-            </local:PlaceholderTextBox>
+            </controls:PlaceholderTextBox>
             <Button Content="&#xE711;" x:Name="DeleteButton" FontFamily="Segoe MDL2 Assets" HorizontalAlignment="Right"
                     Style="{StaticResource FancyButton}"
                     BorderBrush="Transparent" FontSize="20" Background="Transparent" Height="30" Margin="5"
